Refactor unsigned long long ints -> size_t

+++ VolIterator.cpp
@@ -11,12 +11,12 @@
 #include "colours.h"
 
 
-VolIterator::VolIterator(std::string filename, unsigned long long int width, unsigned long long int height, unsigned long long int depth) : width(width), height(height), depth(depth) {
+VolIterator::VolIterator(std::string filename, size_t width, size_t height, size_t depth) : width(width), height(height), depth(depth) {
 	
 	file.open(filename, std::ios::binary);
 }
 
-bool VolIterator::loadSlice(unsigned long long int z) {
+bool VolIterator::loadSlice(size_t z) {
 	assert(z < depth);
 
 	// For now, throw out all slices currently loaded
@@ -52,7 +52,7 @@ void VolIterator::clearSlices() {
 	slices.clear();
 }
 
-VolIterator* VolIterator::Open(std::string filename, unsigned long long int width, unsigned long long int height, unsigned long long int depth) {
+VolIterator* VolIterator::Open(std::string filename, size_t width, size_t height, size_t depth) {
 
 	// Check file path
 	if (!fs::fileExists(filename)) {
@@ -61,8 +61,8 @@ VolIterator* VolIterator::Open(std::string filename, unsigned long long int widt
 	}
 
 	// Check file size
-	unsigned long long int filesize = fs::fileSize(filename);
-	unsigned long long int expected = width * height * depth * sizeof(float);
+	size_t filesize = fs::fileSize(filename);
+	size_t expected = width * height * depth * sizeof(float);
 	if (filesize < expected) {
 		printf(RED "File %s was not found to be the advertised size (should be %llu bytes, found %llu bytes).\n" WHITE, filename.c_str(), expected, filesize);
 		return nullptr;
@@ -72,7 +72,7 @@ VolIterator* VolIterator::Open(std::string filename, unsigned long long int widt
 	return new VolIterator(filename, width, height, depth);
 }
 
-bool VolIterator::exportSlicePng(unsigned long long int z, std::string filename, float minThreshold, float maxThreshold) {
+bool VolIterator::exportSlicePng(size_t z, std::string filename, float minThreshold, float maxThreshold) {
 
 	if (z >= depth) {
 		printf(RED "Invalid slice %llu on volume of size %llu x %llu x %llu.\n" WHITE, z, width, height, depth);
@@ -85,8 +85,8 @@ bool VolIterator::exportSlicePng(unsigned long long int z, std::string filename,
 
 	// Convert slice to 8-bit greyscale image
 	unsigned char* pixels = new unsigned char[width * height];
-	for (unsigned long long int y = 0; y < height; ++y) {
-		for (unsigned long long int x = 0; x < width; ++x) {
+	for (size_t y = 0; y < height; ++y) {
+		for (size_t x = 0; x < width; ++x) {
 			float val = slices[0][y * width + x];
 			val = (val - minThreshold) / (maxThreshold - minThreshold); // 0..1 remap
 			pixels[y * width + x] = val < 0.0f ? 0 : val > 1.0f ? 255 : int(val * 255); // clamp & write
